refactor: changed file structure
Changed file structure to have no code in App.tsx, and changed it to have StopWatch.tsx as the primary parent component
